requirements.txt
streamlit
torch
diffusers
transformers
cuda-python
pillow
